+++ common/src/main/java/com/yanny/ali/plugin/mods/deeper_and_darker/Plugin.java
@@ -1,4 +1,4 @@
-package com.yanny.ali.plugin.mods.moonlight;
+package com.yanny.ali.plugin.mods.deeper_and_darker;
 
 import com.yanny.ali.api.AliEntrypoint;
 import com.yanny.ali.api.IPlugin;
@@ -9,12 +9,11 @@ import com.yanny.ali.plugin.mods.PluginUtils;
 public class Plugin implements IPlugin {
     @Override
     public String getModId() {
-        return "moonlight";
+        return "deeperdarker";
     }
 
     @Override
     public void registerServer(IServerRegistry registry) {
-        PluginUtils.registerConditionTooltip(registry, OptionalPropertyCondition.class);
-        PluginUtils.registerEntry(registry, OptionalItemPool.class);
+        PluginUtils.registerFunctionTooltip(registry, SetPaintingVariantFunction.class);
     }
 }